requirements.txt
playwright >= 1.58.0
behave >= 1.3.3
allure-behave >= 2.15.3
requests >= 2.32.5